requirements.txt
asgiref==3.8.1
beautifulsoup4==4.14.3
better-profanity==0.7.0
bleach==6.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
cryptography==46.0.4
diff-match-patch==20181111
Django==5.2.1
django-allauth==65.11.0
django-ckeditor==6.7.3
django-ckeditor-5==0.2.19
django-crispy-forms==2.5
django-js-asset==3.1.2
djangorestframework==3.16.0
idna==3.11
jwt==1.4.0
pillow==12.1.0
pycparser==3.0
quill-delta==1.0.3
requests==2.32.5
soupsieve==2.8.3
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
webencodings==0.5.1
